requirements_old.txt
amqp==5.1.1
arrow==1.2.3
asgiref==3.7.2
billiard==4.1.0
blessed==1.20.0
branca==0.6.0
celery==5.3.4
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cryptography==41.0.4
defusedxml==0.7.1
Django==4.2.5
django-debug-toolbar==4.2.0
django-environ==0.11.2
django-extensions==3.2.3
django-picklefield==3.1
django-q==1.3.9
django-simple-history==3.4.0
django-sslserver==0.22
easy-thumbnails==2.8.5
folium==0.14.0
geographiclib==2.0
geopy==2.4.0
idna==3.4
Jinja2==3.1.2
kombu==5.3.2
MarkupSafe==2.1.3
numpy==1.26.0
oauthlib==3.2.2
Pillow==10.0.1
prompt-toolkit==3.0.39
pycparser==2.21
PyJWT==2.8.0
pyOpenSSL==23.2.0
pyTelegramBotAPI==4.14.0
python-dateutil==2.8.2
python3-openid==3.2.0
redis==3.5.3
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==5.3.0
social-auth-core==4.4.2
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.0.5
vine==5.0.0
wcwidth==0.2.6
Werkzeug==2.3.7
